refactor(auth): moved AccountInfo component to auth module
